fix small size files

+++ transform.h
@@ -59,7 +59,7 @@ void Highlight_hEdit_right(UINT uStartPos, UINT uEndPos, COLORREF color);
 int readline(PTCHAR buf, long long mark_src);
 void putbyt_buf(PTCHAR buf1, PTCHAR buf2, long long* mark_dst, long long* mark_src);
 void putbyt_addr(int c, PTCHAR Buffer_1, long long* mark_dst);
-void putlong(PTCHAR Buffer_1, int addr, long long* mark_dst, char *first_time);
+void putlong(PTCHAR Buffer_1, long long addr, long long* mark_dst, char *first_time);
 void printline(long long size, long long* mark_dst, long long* mark_src, PTCHAR buf1, PTCHAR buf2);
 void buffer_write_left(PTCHAR Buffer_1, PTCHAR Buffer_2, long long chunk_read, char *first_time);
 void buffer_write_right(PTCHAR Buffer_1, PTCHAR Buffer_2, long long chunk_read, char *first_time);
